requirements.txt
beautifulsoup4>=4.9.2
pymongo>=3.11.1
python-telegram-bot>=12.8.1
requests>=2.24.1
